+++ src/factor_instance_provider/cache/mod.rs
@@ -1,3 +1,21 @@
+mod derivation_path_without_index;
 mod is_pre_derived_keys_cache;
+mod next_derivation_peek_outcome;
+mod pre_derive_keys_cache_key;
+mod unfulfillable_request;
+mod unfulfillable_request_reason;
+mod unfulfillable_requests;
 
+#[cfg(test)]
+mod testing;
+
+pub use derivation_path_without_index::*;
 pub use is_pre_derived_keys_cache::*;
+pub use next_derivation_peek_outcome::*;
+pub use pre_derive_keys_cache_key::*;
+pub use unfulfillable_request::*;
+pub use unfulfillable_request_reason::*;
+pub use unfulfillable_requests::*;
+
+#[cfg(test)]
+pub use testing::*;
