+++ src/factor_instance_provider/cache/next_derivation_peek_outcome.rs
