+++ src/factor_instance_provider/cache/testing/mod.rs
@@ -1,0 +1,5 @@
+#![cfg(test)]
+
+mod in_memory_pre_derived_keys_cache;
+
+pub use in_memory_pre_derived_keys_cache::*;
