+++ src/factor_instance_provider/factor_instance_provider.rs
@@ -6,258 +6,6 @@ use rand::seq::index;
 
 use crate::prelude::*;
 
-/// Used as a map key in `InMemoryPreDerivedKeysCache`.
-#[derive(Clone, PartialEq, Eq, Hash)]
-pub struct PreDeriveKeysCacheKey {
-    factor_source_id: FactorSourceIDFromHash,
-    path_without_index: DerivationPathWithoutIndex,
-}
-
-impl From<HierarchicalDeterministicFactorInstance> for PreDeriveKeysCacheKey {
-    fn from(value: HierarchicalDeterministicFactorInstance) -> Self {
-        Self::new(
-            value.factor_source_id(),
-            DerivationPathWithoutIndex::from(value),
-        )
-    }
-}
-impl PreDeriveKeysCacheKey {
-    pub fn new(
-        factor_source_id: FactorSourceIDFromHash,
-        path_without_index: DerivationPathWithoutIndex,
-    ) -> Self {
-        Self {
-            factor_source_id,
-            path_without_index,
-        }
-    }
-}
-
-/// Like a `DerivationPath` but without the last path component.
-#[derive(Clone, PartialEq, Eq, Hash)]
-pub struct DerivationPathWithoutIndex {
-    network_id: NetworkID,
-    entity_kind: CAP26EntityKind,
-    key_kind: CAP26KeyKind,
-    key_space: KeySpace,
-}
-impl DerivationPathWithoutIndex {
-    fn new(
-        network_id: NetworkID,
-        entity_kind: CAP26EntityKind,
-        key_kind: CAP26KeyKind,
-        key_space: KeySpace,
-    ) -> Self {
-        Self {
-            network_id,
-            entity_kind,
-            key_kind,
-            key_space,
-        }
-    }
-}
-impl From<HierarchicalDeterministicFactorInstance> for DerivationPathWithoutIndex {
-    fn from(value: HierarchicalDeterministicFactorInstance) -> Self {
-        Self::new(
-            value.derivation_path().network_id,
-            value.derivation_path().entity_kind,
-            value.derivation_path().key_kind,
-            value.derivation_path().index.key_space(),
-        )
-    }
-}
-impl From<DerivationPath> for DerivationPathWithoutIndex {
-    fn from(value: DerivationPath) -> Self {
-        Self::new(
-            value.network_id,
-            value.entity_kind,
-            value.key_kind,
-            value.index.key_space(),
-        )
-    }
-}
-
-impl From<(DerivationPathWithoutIndex, HDPathComponent)> for DerivationPath {
-    fn from(value: (DerivationPathWithoutIndex, HDPathComponent)) -> Self {
-        let (without_index, index) = value;
-        assert!(index.is_in_key_space(without_index.key_space));
-        Self::new(
-            without_index.network_id,
-            without_index.entity_kind,
-            without_index.key_kind,
-            index,
-        )
-    }
-}
-
-#[cfg(test)]
-/// A simple `IsPreDerivedKeysCache` which uses in-memory cache instead of on
-/// file which the live implementation will use.
-#[derive(Default)]
-pub struct InMemoryPreDerivedKeysCache {
-    cache:
-        RwLock<HashMap<PreDeriveKeysCacheKey, IndexSet<HierarchicalDeterministicFactorInstance>>>,
-}
-
-impl From<DerivationRequest> for DerivationPathWithoutIndex {
-    fn from(value: DerivationRequest) -> Self {
-        Self::new(
-            value.network_id,
-            value.entity_kind,
-            value.key_kind,
-            value.key_space,
-        )
-    }
-}
-
-#[derive(Clone, PartialEq, Eq, Hash)]
-struct Tuple {
-    request: DerivationRequest,
-    path: DerivationPathWithoutIndex,
-}
-impl Tuple {
-    #[allow(unused)]
-    fn cache_key(&self) -> PreDeriveKeysCacheKey {
-        PreDeriveKeysCacheKey::new(self.request.factor_source_id, self.path.clone())
-    }
-}
-#[cfg(test)]
-impl InMemoryPreDerivedKeysCache {
-    fn tuples(requests: IndexSet<DerivationRequest>) -> IndexSet<Tuple> {
-        requests
-            .clone()
-            .into_iter()
-            .map(|request| Tuple {
-                request,
-                path: DerivationPathWithoutIndex::from(request),
-            })
-            .collect::<IndexSet<Tuple>>()
-    }
-
-    /// Internal helper for implementing `peek`.
-    /// `peek` will will this and map:
-    /// 1. `Err(e)` -> `NextDerivationPeekOutcome::Failure(e)`
-    /// 1. `Ok(None)` -> `NextDerivationPeekOutcome::Fulfillable`
-    /// 1. `Ok(requests)` -> `NextDerivationPeekOutcome::Unfulfillable(UnfulfillableRequests::new(requests))`
-    async fn try_peek(
-        &self,
-        requests: IndexSet<DerivationRequest>,
-    ) -> Result<Option<UnfulfillableRequests>> {
-        let cached = self
-            .cache
-            .try_read()
-            .map_err(|_| CommonError::KeysCacheReadGuard)?;
-
-        let request_and_path_tuples = InMemoryPreDerivedKeysCache::tuples(requests.clone());
-
-        let mut unfulfillable = IndexSet::<UnfulfillableRequest>::new();
-        for tuple in request_and_path_tuples.iter() {
-            let request = tuple.request;
-            let Some(for_key) = cached.get(&tuple.cache_key()) else {
-                unfulfillable.insert(UnfulfillableRequest::empty(request));
-                continue;
-            };
-
-            let factors_left = for_key.len();
-            if factors_left == 0 {
-                unfulfillable.insert(UnfulfillableRequest::empty(request));
-            } else if factors_left == 1 {
-                let last_factor = for_key.last().expect("Just checked length.");
-                unfulfillable.insert(UnfulfillableRequest::last(request, last_factor));
-            } else {
-                // all good
-                continue;
-            }
-        }
-
-        if unfulfillable.is_empty() {
-            Ok(None)
-        } else {
-            Ok(Some(UnfulfillableRequests::new(unfulfillable)))
-        }
-    }
-}
-
-#[cfg(test)]
-#[async_trait::async_trait]
-impl IsPreDerivedKeysCache for InMemoryPreDerivedKeysCache {
-    async fn insert(
-        &self,
-        derived_factors_map: IndexMap<
-            PreDeriveKeysCacheKey,
-            IndexSet<HierarchicalDeterministicFactorInstance>,
-        >,
-    ) -> Result<()> {
-        for (key, values) in derived_factors_map.iter() {
-            for value in values {
-                assert_eq!(
-                    value.factor_source_id(),
-                    key.factor_source_id,
-                    "Discrepancy! FactorSourceID mismatch, this is a developer error."
-                );
-            }
-        }
-
-        let mut write_guard = self
-            .cache
-            .try_write()
-            .map_err(|_| CommonError::KeysCacheWriteGuard)?;
-
-        for (key, derived_factors) in derived_factors_map {
-            if let Some(existing_factors) = write_guard.get_mut(&key) {
-                existing_factors.extend(derived_factors);
-            } else {
-                write_guard.insert(key, derived_factors);
-            }
-        }
-        drop(write_guard);
-
-        Ok(())
-    }
-
-    async fn consume_next_factor_instances(
-        &self,
-        requests: IndexSet<DerivationRequest>,
-    ) -> Result<IndexMap<DerivationRequest, HierarchicalDeterministicFactorInstance>> {
-        let mut cached = self
-            .cache
-            .try_write()
-            .map_err(|_| CommonError::KeysCacheWriteGuard)?;
-
-        let mut instances_read_from_cache =
-            IndexMap::<DerivationRequest, HierarchicalDeterministicFactorInstance>::new();
-
-        let request_and_path_tuples = InMemoryPreDerivedKeysCache::tuples(requests.clone());
-
-        for tuple in request_and_path_tuples {
-            let for_key = cached
-                .get_mut(&tuple.cache_key())
-                .ok_or(CommonError::KeysCacheUnknownKey)?;
-            let read_from_cache = for_key
-                .first()
-                .ok_or(CommonError::KeysCacheEmptyForKey)?
-                .clone();
-            assert!(
-                read_from_cache.matches(&tuple.request),
-                "incorrect implementation"
-            );
-            for_key.shift_remove(&read_from_cache);
-            instances_read_from_cache.insert(tuple.request, read_from_cache);
-        }
-
-        Ok(instances_read_from_cache)
-    }
-
-    async fn peek(&self, requests: IndexSet<DerivationRequest>) -> NextDerivationPeekOutcome {
-        let outcome = self.try_peek(requests).await;
-        match outcome {
-            Ok(None) => NextDerivationPeekOutcome::Fulfillable,
-            Ok(Some(unfulfillable)) => NextDerivationPeekOutcome::Unfulfillable(unfulfillable),
-            Err(e) => NextDerivationPeekOutcome::Failure(e),
-        }
-    }
-}
-
 /// A coordinator of sorts between `PreDeriveKeysCache`, `KeysCollector` and Gateway,
 /// used to provide `HierarchicalDeterministicFactorInstance`s for a given `DerivationRequest`.
 ///
@@ -269,6 +17,7 @@ pub struct FactorInstanceProvider {
     derivation_interactors: Arc<dyn KeysDerivationInteractors>,
     cache: Arc<dyn IsPreDerivedKeysCache>,
 }
+
 impl FactorInstanceProvider {
     pub fn new(
         gateway: Arc<dyn Gateway>,
@@ -283,65 +32,69 @@ impl FactorInstanceProvider {
     }
 }
 
-impl From<(DerivationRequest, HDPathComponent)> for DerivationPath {
-    fn from(value: (DerivationRequest, HDPathComponent)) -> Self {
-        let (request, index) = value;
-        DerivationPath::new(
-            request.network_id,
-            request.entity_kind,
-            request.key_kind,
-            index,
-        )
-    }
-}
-
-impl MatrixOfFactorInstances {
-    fn highest_derivation_path_index(
+// ===== ********** =====
+// ===== PUBLIC API =====
+// ===== ********** =====
+impl FactorInstanceProvider {
+    pub async fn provide_genesis_factor_for<'p>(
         &self,
-        request: &DerivationRequest,
-    ) -> Option<HDPathComponent> {
-        self.all_factors()
-            .into_iter()
-            .filter(|f| f.matches(request))
-            .map(|f| f.derivation_path().index)
-            .max()
-    }
-}
-impl SecurifiedEntityControl {
-    fn highest_derivation_path_index(
-        &self,
-        request: &DerivationRequest,
-    ) -> Option<HDPathComponent> {
-        self.matrix.highest_derivation_path_index(request)
-    }
-}
-impl SecurifiedEntity {
-    fn highest_derivation_path_index(
-        &self,
-        request: &DerivationRequest,
-    ) -> Option<HDPathComponent> {
-        self.control.highest_derivation_path_index(request)
-    }
-}
+        factor_source_id: FactorSourceIDFromHash,
+        entity_kind: CAP26EntityKind,
+        network_id: NetworkID,
+        profile: &'p Profile,
+    ) -> Result<HierarchicalDeterministicFactorInstance> {
+        let key_kind = CAP26KeyKind::TransactionSigning;
 
-impl From<HierarchicalDeterministicFactorInstance> for DerivationRequest {
-    fn from(value: HierarchicalDeterministicFactorInstance) -> Self {
-        let key_space = value.derivation_path().index.key_space();
-        let key_kind = value.derivation_path().key_kind;
-        let entity_kind = value.derivation_path().entity_kind;
-        let network_id = value.derivation_path().network_id;
-        let factor_source_id = value.factor_source_id();
-
-        Self::new(
-            key_space,
+        let request = DerivationRequest::new(
+            KeySpace::Unsecurified,
             entity_kind,
             key_kind,
             factor_source_id,
             network_id,
-        )
+        );
+
+        let derived_factors_map = self
+            .provide_factor_instances(profile, IndexSet::just(request))
+            .await?;
+
+        let derived_factor = derived_factors_map
+            .into_iter()
+            .next()
+            .ok_or(CommonError::InstanceProviderFailedToCreateGenesisFactor)?
+            .1;
+
+        Ok(derived_factor)
+    }
+
+    pub async fn provide_factor_instances<'p>(
+        &self,
+        profile: &'p Profile,
+        requests: IndexSet<DerivationRequest>,
+    ) -> Result<IndexMap<DerivationRequest, HierarchicalDeterministicFactorInstance>> {
+        let peek_outcome = self.cache.peek(requests.clone()).await;
+
+        match peek_outcome {
+            NextDerivationPeekOutcome::Failure(e) => {
+                error!("Failed to peek next derivation index: {:?}", e);
+                Err(e)
+            }
+            NextDerivationPeekOutcome::Fulfillable => {
+                self.cache.consume_next_factor_instances(requests).await
+            }
+            NextDerivationPeekOutcome::Unfulfillable(unfulfillable) => {
+                let fulfillable = requests
+                    .difference(&unfulfillable.requests())
+                    .cloned()
+                    .collect::<IndexSet<_>>();
+                self.fulfill(profile, fulfillable, unfulfillable).await
+            }
+        }
     }
 }
 
+// ===== ********** =====
+// ===== PRIVATE API =====
+// ===== ********** =====
 impl FactorInstanceProvider {
     async fn derive_new_and_fill_cache<'p>(
         &self,
@@ -537,60 +290,65 @@ impl FactorInstanceProvider {
     }
 }
 
-impl FactorInstanceProvider {
-    pub async fn provide_genesis_factor_for<'p>(
-        &self,
-        factor_source_id: FactorSourceIDFromHash,
-        entity_kind: CAP26EntityKind,
-        network_id: NetworkID,
-        profile: &'p Profile,
-    ) -> Result<HierarchicalDeterministicFactorInstance> {
-        let key_kind = CAP26KeyKind::TransactionSigning;
+// ===== ********** =====
+// =====  HELPERS   =====
+// ===== ********** =====
+impl From<(DerivationRequest, HDPathComponent)> for DerivationPath {
+    fn from(value: (DerivationRequest, HDPathComponent)) -> Self {
+        let (request, index) = value;
+        DerivationPath::new(
+            request.network_id,
+            request.entity_kind,
+            request.key_kind,
+            index,
+        )
+    }
+}
 
-        let request = DerivationRequest::new(
-            KeySpace::Unsecurified,
+impl MatrixOfFactorInstances {
+    fn highest_derivation_path_index(
+        &self,
+        request: &DerivationRequest,
+    ) -> Option<HDPathComponent> {
+        self.all_factors()
+            .into_iter()
+            .filter(|f| f.matches(request))
+            .map(|f| f.derivation_path().index)
+            .max()
+    }
+}
+impl SecurifiedEntityControl {
+    fn highest_derivation_path_index(
+        &self,
+        request: &DerivationRequest,
+    ) -> Option<HDPathComponent> {
+        self.matrix.highest_derivation_path_index(request)
+    }
+}
+impl SecurifiedEntity {
+    fn highest_derivation_path_index(
+        &self,
+        request: &DerivationRequest,
+    ) -> Option<HDPathComponent> {
+        self.control.highest_derivation_path_index(request)
+    }
+}
+
+impl From<HierarchicalDeterministicFactorInstance> for DerivationRequest {
+    fn from(value: HierarchicalDeterministicFactorInstance) -> Self {
+        let key_space = value.derivation_path().index.key_space();
+        let key_kind = value.derivation_path().key_kind;
+        let entity_kind = value.derivation_path().entity_kind;
+        let network_id = value.derivation_path().network_id;
+        let factor_source_id = value.factor_source_id();
+
+        Self::new(
+            key_space,
             entity_kind,
             key_kind,
             factor_source_id,
             network_id,
-        );
-
-        let derived_factors_map = self
-            .provide_factor_instances(profile, IndexSet::just(request))
-            .await?;
-
-        let derived_factor = derived_factors_map
-            .into_iter()
-            .next()
-            .ok_or(CommonError::InstanceProviderFailedToCreateGenesisFactor)?
-            .1;
-
-        Ok(derived_factor)
-    }
-
-    pub async fn provide_factor_instances<'p>(
-        &self,
-        profile: &'p Profile,
-        requests: IndexSet<DerivationRequest>,
-    ) -> Result<IndexMap<DerivationRequest, HierarchicalDeterministicFactorInstance>> {
-        let peek_outcome = self.cache.peek(requests.clone()).await;
-
-        match peek_outcome {
-            NextDerivationPeekOutcome::Failure(e) => {
-                error!("Failed to peek next derivation index: {:?}", e);
-                Err(e)
-            }
-            NextDerivationPeekOutcome::Fulfillable => {
-                self.cache.consume_next_factor_instances(requests).await
-            }
-            NextDerivationPeekOutcome::Unfulfillable(unfulfillable) => {
-                let fulfillable = requests
-                    .difference(&unfulfillable.requests())
-                    .cloned()
-                    .collect::<IndexSet<_>>();
-                self.fulfill(profile, fulfillable, unfulfillable).await
-            }
-        }
+        )
     }
 }
