+++ src/factor_instance_provider/mod.rs
@@ -4,20 +4,12 @@ mod factor_instance_provider;
 
 mod cache;
 mod gateway;
-mod next_derivation_peek_outcome;
 mod profile_extensions;
 mod securify;
-mod unfulfillable_request;
-mod unfulfillable_request_reason;
-mod unfulfillable_requests;
 
 pub use cache::*;
 pub use derivation_request::*;
 pub use factor_instance_provider::*;
 pub use gateway::*;
-pub use next_derivation_peek_outcome::*;
 pub use profile_extensions::*;
 pub use securify::*;
-pub use unfulfillable_request::*;
-pub use unfulfillable_request_reason::*;
-pub use unfulfillable_requests::*;
